// AdminFrontend/src/app/Model/User.model.ts
export class User {
    userId: Number;
    name: String;
    username: String;
    password: String;
    email: String;
    role: String;
    isActive;
    securityQuestion: String;
    securityAnswer: String;
    phoneNumber: String;
    alternatePhoneNumber: String;
    alternateEmail: String;
    address: String;
    isApproved: boolean;
    rating: number;
    cCF = [];
    cPF = [];
    orders = [];
    cC = [];
    addresses = [];
    products = [];
    pF = [];
    coupons = [];
  }